+++ src/components/forms/FormNumberInput.ts
@@ -1,0 +1,20 @@
+//
+// Imports
+//
+
+import { DE, InputElementAttributes } from "@donutteam/document-builder";
+
+//
+// Exports
+//
+
+export function FormNumberInput(attributes : InputElementAttributes) : DE
+{
+	return new DE("input",
+		{
+			class: "component-form-number-input",
+			type: "number",
+
+			...attributes,
+		});
+}