+++ src/index.ts
@@ -7,6 +7,7 @@ export * from "./components/forms/FormCheckboxInput.js";
 export * from "./components/forms/FormEmailInput.js";
 export * from "./components/forms/FormFieldset.js";
 export * from "./components/forms/FormNoticeContainer.js";
+export * from "./components/forms/FormNumberInput.js";
 export * from "./components/forms/FormRadioInput.js";
 export * from "./components/forms/FormTextInput.js";